+++ openshift/wsgi.py
@@ -28,11 +28,8 @@ sys.path.append(os.path.join(os.environ['OPENSHIFT_REPO_DIR'], 'openshift'))
 
 os.environ['DJANGO_SETTINGS_MODULE'] = 'weblate.settings_openshift'
 
-virtenv = os.environ['OPENSHIFT_PYTHON_DIR'] + '/virtenv/'
-virtualenv = os.path.join(virtenv, 'bin/activate_this.py')
-try:
-    execfile(virtualenv, dict(__file__=virtualenv))
-except IOError:
-    pass
+execfile(os.path.join(
+    os.environ['OPENSHIFT_REPO_DIR'], 'openshift', 'virtualenv.py'
+))
 
 application = get_wsgi_application()